refactored backend code
