+++ back-end/recipeZ/src/main/java/com/mathewtri/recipeZ/controller/UserController.java
@@ -1,7 +1,7 @@
 package com.mathewtri.recipeZ.controller;
 
 import com.mathewtri.recipeZ.model.User;
-import com.mathewtri.recipeZ.service.IUserService;
+import com.mathewtri.recipeZ.service.user.IUserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -31,72 +31,59 @@ public class UserController {
 
     // signup
     @PostMapping("/users/create")
-    public ResponseEntity<Boolean> createUser(@RequestBody User user){
+    public ResponseEntity<Boolean> createUser(@RequestBody User user) {
         boolean success = userService.createUser(user);
-        if(success){
+        if (success) {
             return ResponseEntity.ok(true);
-        }else{
+        } else {
             return ResponseEntity.ok(false);
         }
     }
 
     // login
     @PostMapping("/users")
-    public ResponseEntity<User> loginUser(@RequestBody User user){
+    public ResponseEntity<User> loginUser(@RequestBody User user) {
         User fetchedUser = userService.fetchUserByEmailAndPassword(user.getEmail(), user.getPassword());
         System.out.println(fetchedUser);
-        if(fetchedUser == null)
-        {
+        if (fetchedUser == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        }
-        else
-        {
+        } else {
             return ResponseEntity.ok(fetchedUser);
         }
     }
 
     @GetMapping("/users")
-    public ResponseEntity<List<User>> fetchUsers(){
+    public ResponseEntity<List<User>> fetchUsers() {
         List<User> users = userService.fetchUsers();
         return ResponseEntity.ok(users);
     }
 
     @GetMapping("/users/{userId}")
-    public ResponseEntity<User> fetchUserById(@PathVariable String userId){
+    public ResponseEntity<User> fetchUserById(@PathVariable String userId) {
         User user = userService.fetchUserById(userId);
-        if(user == null)
-        {
+        if (user == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        }
-        else
-        {
+        } else {
             return ResponseEntity.ok(user);
         }
     }
 
     @GetMapping("/users/search")
-    public ResponseEntity<User> fetchUserByEmail(@RequestParam String email){
+    public ResponseEntity<User> fetchUserByEmail(@RequestParam String email) {
         User user = userService.fetchUserByEmail(email);
-        if(user == null)
-        {
+        if (user == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        }
-        else
-        {
+        } else {
             return ResponseEntity.ok(user);
         }
     }
 
     @DeleteMapping("/users/{userId}")
-    public String delete(@PathVariable String userId)
-    {
+    public String delete(@PathVariable String userId) {
         User user = userService.fetchUserById(userId);
-        if(user == null)
-        {
+        if (user == null) {
             return "Failed to Delete User: id not found";
-        }
-        else
-        {
+        } else {
             userService.deleteUser(userId);
             return "Deleted Student with id " + userId;
         }
