--- back-end/recipeZ/src/main/java/com/mathewtri/recipeZ/filter/CustomAuthenticationFilter.java
@@ -1,71 +1,0 @@
-//package com.mathewtri.recipeZ.filter;
-//
-//import com.auth0.jwt.JWT;
-//import com.auth0.jwt.algorithms.Algorithm;
-//import com.fasterxml.jackson.databind.ObjectMapper;
-//import lombok.extern.slf4j.Slf4j;
-//import org.springframework.security.authentication.AuthenticationManager;
-//import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-//import org.springframework.security.core.Authentication;
-//import org.springframework.security.core.AuthenticationException;
-//import org.springframework.security.core.GrantedAuthority;
-//import org.springframework.security.core.userdetails.User;
-//import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-//
-//import javax.servlet.FilterChain;
-//import javax.servlet.ServletException;
-//import javax.servlet.http.HttpServletRequest;
-//import javax.servlet.http.HttpServletResponse;
-//import java.io.IOException;
-//import java.util.Date;
-//import java.util.HashMap;
-//import java.util.Map;
-//import java.util.stream.Collectors;
-//
-//import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-//
-//@Slf4j
-//public class CustomAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
-//
-//    private final AuthenticationManager authenticationManager;
-//
-//    public CustomAuthenticationFilter (AuthenticationManager authenticationManager){
-//        this.authenticationManager = authenticationManager;
-//    }
-//
-//    @Override
-//    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException {
-//        System.out.println(request.getRequestURL());
-//        String username = request.getParameter("username");
-//        String password = request.getParameter("password");
-//        log.info("Username is: {}", username); log.info("Password is: {}", password);
-//        UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(username, password);
-//        return authenticationManager.authenticate(authenticationToken);
-//    }
-//
-//    @Override
-//    protected void successfulAuthentication(HttpServletRequest request, HttpServletResponse response, FilterChain chain, Authentication authentication) throws IOException, ServletException {
-//        User user = (User) authentication.getPrincipal();
-//        Algorithm algorithm = Algorithm.HMAC256("secret".getBytes());
-//        String accessToken = JWT.create()
-//                .withSubject(user.getUsername())
-//                .withExpiresAt(new Date(System.currentTimeMillis() + 10 * 60 * 1000))
-//                .withIssuer(request.getRequestURL().toString()) // name of organization
-//                .withClaim("roles", user.getAuthorities()
-//                        .stream()
-//                        .map(GrantedAuthority::getAuthority).collect(Collectors.toList()))
-//                .sign(algorithm);
-//
-//        String refreshToken = JWT.create()
-//                .withSubject(user.getUsername())
-//                .withExpiresAt(new Date(System.currentTimeMillis() + 30 * 60 * 1000))
-//                .withIssuer(request.getRequestURL().toString()) // name of organization
-//                .sign(algorithm);
-//
-//        Map<String, String> tokens = new HashMap<>();
-//        tokens.put("access_token", accessToken);
-//        tokens.put("refresh_token", refreshToken);
-//        response.setContentType(APPLICATION_JSON_VALUE);
-//        new ObjectMapper().writeValue(response.getOutputStream(), tokens);
-//    }
-//}
