+++ back-end/recipeZ/src/main/java/com/mathewtri/recipeZ/repository/recipe/IRecipeRepository.java
@@ -1,4 +1,4 @@
-package com.mathewtri.recipeZ.service;
+package com.mathewtri.recipeZ.repository.recipe;
 
 import com.mathewtri.recipeZ.model.Ingredient;
 import com.mathewtri.recipeZ.model.Recipe;
@@ -6,8 +6,7 @@ import com.mathewtri.recipeZ.model.User;
 
 import java.util.List;
 
-public interface IRecipeService {
-
+public interface IRecipeRepository {
     boolean createRecipe(String userId, Recipe recipe);
     List<Recipe> fetchRecipes(String userId);
     Recipe fetchRecipeById(String userId, String recipeId);
