+++ back-end/recipeZ/src/main/java/com/mathewtri/recipeZ/repository/user/UserRepository.java
@@ -1,4 +1,4 @@
-package com.mathewtri.recipeZ.repository;
+package com.mathewtri.recipeZ.repository.user;
 
 import com.mathewtri.recipeZ.model.User;
 import org.springframework.beans.factory.annotation.Autowired;
