+++ back-end/recipeZ/src/main/java/com/mathewtri/recipeZ/service/recipe/IRecipeService.java
@@ -1,12 +1,11 @@
-package com.mathewtri.recipeZ.repository;
+package com.mathewtri.recipeZ.service.recipe;
 
-import com.mathewtri.recipeZ.model.Ingredient;
 import com.mathewtri.recipeZ.model.Recipe;
-import com.mathewtri.recipeZ.model.User;
 
 import java.util.List;
 
-public interface IRecipeRepository {
+public interface IRecipeService {
+
     boolean createRecipe(String userId, Recipe recipe);
     List<Recipe> fetchRecipes(String userId);
     Recipe fetchRecipeById(String userId, String recipeId);
